[Refactor] Add Base module
