+++ be/src/base/compiler_util.h
@@ -14,5 +14,6 @@
 
 #pragma once
 
-// Backward-compatible shim. Canonical definitions live in base.
-#include "base/statusor_internal.h"
+// Backward-compatible shim. Canonical definitions live in gutil to avoid
+// introducing BE module dependencies into gutil.
+#include "gutil/compiler_util.h"
