+++ be/src/base/statusor.cpp
@@ -11,13 +11,13 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#include "common/statusor.h"
+#include "base/statusor.h"
 
 #include <cstdlib>
 #include <iostream>
 #include <utility>
 
-#include "common/status.h"
+#include "base/status.h"
 
 namespace starrocks {
