+++ be/src/common/logging.h
@@ -13,25 +13,6 @@
 // limitations under the License.
 
 #pragma once
-// glog MUST be included before gflags. Instead of including them,
-// our files should include this file instead.
-#include <glog/logging.h>
 
-// Define VLOG levels.  We want display per-row info less than per-file which
-// is less than per-query.  For now per-connection is the same as per-query.
-#define VLOG_CONNECTION VLOG(2)
-#define VLOG_RPC VLOG(8)
-#define VLOG_QUERY VLOG(2)
-#define VLOG_FILE VLOG(2)
-#define VLOG_OPERATOR VLOG(3)
-#define VLOG_ROW VLOG(10)
-#define VLOG_PROGRESS VLOG(2)
-#define VLOG_CACHE VLOG(3)
-
-#define VLOG_CONNECTION_IS_ON VLOG_IS_ON(2)
-#define VLOG_RPC_IS_ON VLOG_IS_ON(2)
-#define VLOG_QUERY_IS_ON VLOG_IS_ON(2)
-#define VLOG_FILE_IS_ON VLOG_IS_ON(2)
-#define VLOG_OPERATOR_IS_ON VLOG_IS_ON(3)
-#define VLOG_ROW_IS_ON VLOG_IS_ON(10)
-#define VLOG_PROGRESS_IS_ON VLOG_IS_ON(2)
+// Backward-compatible shim. Canonical definitions live in base.
+#include "base/logging.h"
