+++ be/src/starrocks_format/CMakeLists.txt
@@ -22,6 +22,7 @@ set(EXECUTABLE_OUTPUT_PATH "${BUILD_DIR}/src/starrocks_format")
 # Set starrocks libraries
 set(STARROCKS_LIBS
     ${WL_START_GROUP}
+    Base
     Common
     Column
     Connector
